// sdk/db/scripts/dapp_interface.ts
export interface dappInfo {
    id: string
    name: string
    address: any
    symbol: string
    url: string
    description: string
    chain: string
    logo: string
    audits: string
    audit_note: any
    gecko_id: any
    cmcId: any
    category: string
    chains: string[]
    module: string
    twitter: string
    forkedFrom: any[]
    oracles: any[]
    listedAt: number
    slug: string
    tvl: number
    change_1h: number
    change_1d: number
    change_7d: number
  }